Update on "Rename TorchScript compiler to IR emitter to better reflect its function."
Differential Revision: [D19806503](https://our.internmc.facebook.com/intern/diff/D19806503)

[ghstack-poisoned]
