+++ test/cpp/jit/test_alias_analysis.cpp
@@ -3,7 +3,7 @@
 #include "test/cpp/jit/test_base.h"
 #include "torch/csrc/jit/custom_operator.h"
 #include "torch/csrc/jit/passes/alias_analysis.h"
-#include "torch/csrc/jit/script/iremitter.h"
+#include "torch/csrc/jit/script/ir_emitter.h"
 #include "torch/csrc/utils/memory.h"
 
 namespace torch {
