+++ test/cpp/jit/test_fuser.cpp
@@ -35,7 +35,7 @@
 #include <torch/csrc/jit/testing/file_check.h>
 #include "ATen/core/ivalue.h"
 #include "torch/csrc/jit/graph_executor.h"
-#include "torch/csrc/jit/script/iremitter.h"
+#include "torch/csrc/jit/script/ir_emitter.h"
 #include "torch/csrc/jit/script/module.h"
 
 #include "onnx/onnx_pb.h"
