+++ torch/csrc/jit/autodiff.cpp
@@ -9,7 +9,7 @@
 #include <torch/csrc/jit/passes/dead_code_elimination.h>
 #include <torch/csrc/jit/passes/inliner.h>
 #include <torch/csrc/jit/passes/lower_tuples.h>
-#include <torch/csrc/jit/script/iremitter.h>
+#include <torch/csrc/jit/script/ir_emitter.h>
 #include <torch/csrc/jit/symbolic_script.h>
 
 #include <algorithm>
