+++ torch/csrc/jit/docs/OVERVIEW.md
@@ -510,9 +510,9 @@ The Parser is written as a [top-down precedence parser](https://eli.thegreenplac
 
 ## IR Emitter ##
 
-[script/iremitter.h](../script/iremitter.h)
+[script/ir_emitter.h](../script/ir_emitter.h)
 
-The file iremitter.cpp translates Trees into Modules. The main entrypoint is `defineMethodsInModule` which takes a list of Def Tree Views representing function definitions and adds them as Methods to the module. During the lowering processing _semantic checking_ occurs. The IR emitter checks that all used variables are defined (sometimes called scope checking), and that all values have compatible types (type-checking). During this process it also emits the graph nodes corresponding to each statement in the Tree and generates a FunctionSchema for the whole definition.
+The file ir_emitter.cpp translates Trees into Modules. The main entrypoint is `defineMethodsInModule` which takes a list of Def Tree Views representing function definitions and adds them as Methods to the module. During the lowering processing _semantic checking_ occurs. The IR emitter checks that all used variables are defined (sometimes called scope checking), and that all values have compatible types (type-checking). During this process it also emits the graph nodes corresponding to each statement in the Tree and generates a FunctionSchema for the whole definition.
 
 A few helper objects exist in the lowering process.  SugaredValues are special values that represent objects that can appear during compilation but that are not first class values. For instance, in TorchScript methods `self` refers to the module, and `self.weight` refers to a Parameter of the module. Neither are first-class Types and have no corresponding Value in a graph. Resolver objects are std::functions that resolve externally-defined variables to SugaredValues. For instance, the identifier `torch` which contains most of our built-in ops is looked up through Resolver objects which interact with the python state of the program.
 
@@ -530,7 +530,7 @@ Finally, normal Values are also represented by the SimpleValue SugaredValue in p
 
 ## Resolver ##
 
-[script/iremitter.h](../script/iremitter.h)
+[script/ir_emitter.h](../script/ir_emitter.h)
 
 Any undefined variable during compilation is resolved with a call to an externally-provided Resolver. When called from Python (e.g `torch.jit.script`) this resolver interacts with the Python runtime via pybind11 to resolve symbols like `torch` and `math` to their Python equivalents.
 
@@ -540,7 +540,7 @@ This makes it possible to use most of the IR emitter functionality when python i
 
 ## Environment ##
 
-[script/iremitter.cpp](../script/iremitter.cpp)
+[script/ir_emitter.cpp](../script/ir_emitter.cpp)
 
 The Environment object tracks the assignment of variable names during compilation. It is local to the IR emitter file. A stack of environments exist, with a new environment being created for sub-blocks introduced by control flow. The Environment keeps two tables, one for values which are not first class in the type system (Sugared values) and a type table for values which are. When first class values are set, we emit a prim::Store, and when they are referenced we emit a prim::Load. Sugared values are not re-assignable. The graph is converted to SSA in the convertToSSA pass.
