+++ torch/csrc/jit/symbolic_script.cpp
@@ -1,6 +1,6 @@
 #include <torch/csrc/jit/symbolic_script.h>
 #include <torch/csrc/jit/operator.h>
-#include <torch/csrc/jit/script/iremitter.h>
+#include <torch/csrc/jit/script/ir_emitter.h>
 
 namespace torch {
 namespace jit {
